+++ libs/langgraph/langgraph/pregel/write.py
@@ -14,7 +14,7 @@ from typing import (
 from langchain_core.runnables import Runnable, RunnableConfig
 from langchain_core.runnables.utils import ConfigurableFieldSpec
 
-from langgraph.constants import CONF, CONFIG_KEY_SEND, TASKS, Send
+from langgraph.constants import CONF, CONFIG_KEY_SEND, MISSING, TASKS, Send
 from langgraph.errors import InvalidUpdateError
 from langgraph.utils.runnable import RunnableCallable
 
@@ -41,6 +41,8 @@ class ChannelWriteTupleEntry(NamedTuple):
     """Function to extract tuples from value."""
     value: Any = PASSTHROUGH
     """Value to write, or PASSTHROUGH to use the input."""
+    declared: Optional[Sequence[tuple[str, Any]]] = None
+    """Optional, declared writes for static analysis."""
 
 
 class ChannelWrite(RunnableCallable):
@@ -121,6 +123,7 @@ class ChannelWrite(RunnableCallable):
     def do_write(
         config: RunnableConfig,
         writes: Sequence[Union[ChannelWriteEntry, ChannelWriteTupleEntry, Send]],
+        allow_passthrough: bool = True,
         require_at_least_one_of: Optional[Sequence[str]] = None,  # ignored
     ) -> None:
         # validate
@@ -130,10 +133,10 @@ class ChannelWrite(RunnableCallable):
                     raise InvalidUpdateError(
                         "Cannot write to the reserved channel TASKS"
                     )
-                if w.value is PASSTHROUGH:
+                if w.value is PASSTHROUGH and not allow_passthrough:
                     raise InvalidUpdateError("PASSTHROUGH value must be replaced")
             if isinstance(w, ChannelWriteTupleEntry):
-                if w.value is PASSTHROUGH:
+                if w.value is PASSTHROUGH and not allow_passthrough:
                     raise InvalidUpdateError("PASSTHROUGH value must be replaced")
         # assemble writes
         tuples: list[tuple[str, Any]] = []
@@ -162,14 +165,26 @@ class ChannelWrite(RunnableCallable):
         """Used by PregelNode to distinguish between writers and other runnables."""
         return (
             isinstance(runnable, ChannelWrite)
-            or getattr(runnable, "_is_channel_writer", False) is True
+            or getattr(runnable, "_is_channel_writer", MISSING) is not MISSING
         )
 
     @staticmethod
-    def register_writer(runnable: R) -> R:
+    def get_declared_writes(
+        runnable: Runnable,
+    ) -> Optional[Sequence[Union[ChannelWriteEntry, Send]]]:
+        """Used to get the writes a writer declares for static analysis."""
+        if writes := getattr(runnable, "_is_channel_writer", MISSING):
+            return writes if writes is not MISSING else None
+
+    @staticmethod
+    def register_writer(
+        runnable: R,
+        declared: Optional[Sequence[Union[ChannelWriteEntry, Send]]] = None,
+    ) -> R:
         """Used to mark a runnable as a writer, so that it can be detected by is_writer.
-        Instances of ChannelWrite are automatically marked as writers."""
+        Instances of ChannelWrite are automatically marked as writers.
+        Optionally, a list of declared writes can be passed for static analysis."""
         # using object.__setattr__ to work around objects that override __setattr__
         # eg. pydantic models and dataclasses
-        object.__setattr__(runnable, "_is_channel_writer", True)
+        object.__setattr__(runnable, "_is_channel_writer", declared)
         return runnable
